changed cinfiguration from x86 to x64

+++ Examples/IntegerSet/IntegerSet_02_Vector.cpp
@@ -20,7 +20,7 @@ namespace IntegerSetVector {
 
     // getter
     int IntegerSet::size() const {
-        return m_set.size();
+        return static_cast<int> (m_set.size());
     }
 
     bool IntegerSet::isEmpty() const {
